+++ include/vyb/vre/llvm/codegen.hpp
@@ -395,6 +395,8 @@ private:
 
     // String type methods
     void handleStringMethod(vyb::ast::CallExpression* node, const std::string& objectName, const std::string& methodName);
+    void handleStringMethodOnValue(vyb::ast::CallExpression* node, llvm::Value* strPtr, const std::string& methodName);
+    void dispatchStringMethod(vyb::ast::CallExpression* node, llvm::Value* strPtr, llvm::Type* strStructType, const std::string& methodName);
     void handleStringLen(vyb::ast::CallExpression* node, llvm::Value* strPtr, llvm::Type* strStructType);
     void handleStringConcat(vyb::ast::CallExpression* node, llvm::Value* strPtr, llvm::Type* strStructType);
     void handleStringSubstring(vyb::ast::CallExpression* node, llvm::Value* strPtr, llvm::Type* strStructType);